refactor: Separate assistance cache
The assistance data is used readonly during loading of the timetables.
To enhance access performance, a cache is implemented for lookup
based on the location.
Since the locations between timetables are mostly disjoint, i.e. the
cache hits are confined to each separate timetable, the cache performance
doesn't suffer when separating the caches into one for each timetable.

To minimise resource consumption, the caches reuse the same underlying
data structures in case of a cache miss. This isn't a problem as the
data is used readonly after initialisation. To load the data, use the
assistance_times_data struct which then gets passed to each
assistance_times to create the cache.

This refactoring reduces the interdependence between the loading
processes.