Refactor: Reconstruct bitfield map from timetable
To efficiently store bitfields, they are used and reused across
timetables. A hash_map maintains this mapping to allow quick checking
of existing bitfields.
After loading, this hash_map isn't needed anymore, so the timetable
only stores the mapping from indices to bitfields (as a vector).
From the timetable, it is possible to reconstruct the hash_map. This
eliminates another global variable.

For now, only assert that the reconstruction and the original are
equal.