feat: Add change detector for loading
The change detector is supposed to identify sources that have changed
or rather have not changed since the last run. To be able to make
that determination, it checks that the filepath is the same and the
file hasn't been written to since the last run. This then implies
that the cache file is still valid.
As such, save that information for later use.

The timestamp returned by fs::last_write_time is relative to some
filesystem clock which may or may not change at any time. In order
to reliably compare those timestamps, they are converted to the
system time which represents time as the Unix epoch (since C++20,
previously not defined but very likely).
To then store the value, it is converted into a 64bit int representing
the nanoseconds since the Unix epoch.