feat: Use cached timetables to improve import time
Since the intermediate data necessary for caching is stored, we
can use it to load from a previous checkpoint when restarting an import.

To find out which sources we can skip, we load the previous change
detector and compare it with the current one. Due to the linear
dependency of importing, we can just check that each source in the list
is unchanged in the order they are presented.
If a changed source is found (path and/or last write time don't match),
all successive sources will also need recomputing, thus break.

In order to restart from cache, we need to load the cached result.
Due to linearity, again, we can just load the result of the previous
source which hasn't changed and thus would produce equivalent
timetable and shape data.
There could still be issues though. If the cached timetable cannot be
loaded for some reason, we can't use it and have to recompute it
anyway. The same is true if the date_range doesn't match. In that case,
the then previous one has to be loaded (and the one before that, etc.).
If the very first source has to be recomputed, there is no cache left,
and the initial timetable with the special stations is created and
provided.

The timetable is always finalised at the end, so changing finalize_opt
don't have to be considered. If the assistance_times change, the cache
may be invalid and has to be deleted. The code doesn't check for that.

It's probably possible reorder the input sources to potentially get
more in common with the cached state, but that would make equivalence
a bit harder to prove. The order usually doesn't change anyway or at
least it can be prevented on the caller side.