feat(bitfields_): Use empty bitfields and merge later
The bitfield indices directly impact the timetable in various ways. This
means that the resulting indices need to be adjusted and the references
corrected.

To that end, affected vectors of the original timetable are saved, before
the bitfields_ member is cleared. We can then load the file with an empty
hashmap.

When the file has been loaded, the new traffic_days are calculated. This
is necessary as the old ones use the correct indices, the new ones however
don't and thus need to be adjusted.
Later on, this won't be necessary as such when an empty timetable is passed
to the loader.

Similarly, the old data is restored where relevant to better show the merge.

Finally, merge the new bitfields into the existing timetable and store the
adjusted indices as they might be non-linear. This mapping is then used
to correct the bitfield indices for the traffic days.

No behaviour should change due to this commit.