fix: Account for non-empty shapes_storage
The shape_idx and shape_offset_idx are relative to the shape storage,
so adjust them when merging.
If the indices are invalid, simply adding them together leads to an
overflow which then produces a valid index. The value however is
wrong and can lead to crashes or silent issues.
If a value is invalid, the adjusted value is also invalid.