refactor(trip_direction_string_idx_t): Use empty and merge later
trip_direction_string_idx_t is part of the trip_direction_t variant, which
can be either this type or a location_idx_t. Since both offsets are independent
from one another, it has to be decided at runtime, which offset needs to be
used.
This commit introduces the index_mapping that maps an existing index (of the
new timetable) to the index in the final timetable. This mapping is created
once from the timetable data and from then on read-only. The mapping does not
consume or otherwise change the provided value.